+++ src/model/polling-response-error.ts
@@ -4,7 +4,7 @@
  * @author Christopher Holomek
  * @since 06.01.2020
  */
-export interface PollingResponseError {
+export interface PollingResponseError extends Error {
     code: number,
     message: string,
 }